requirements.txt
Flask==1.0.3
tensorflow==1.13.1
pip==19.1.1
Werkzeug==0.15.4
matplotlib==3.0.3
Keras==2.2.4
scikit_image==0.15.0
flask_restplus==0.12.1
opencv_python==4.1.0.25
numpy==1.16.4
setuptools==41.0.1
h5py==2.9.0
scipy==1.2.1
Pillow==6.0.0
protobuf==3.8.0
imgaug==0.2.9
mrcnn==0.2
seaborn==0.9.0
scikit-image==0.15.0
scikit_learn==0.21.2
pyzmq==18.0.1
